app/requirements.txt
black==22.10.0
Django==4.1.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
drf-yasg==1.21.4
flake8==5.0.4
isort==5.10.1
psycopg2-binary==2.9.4
pytest==7.2.0
pytest-cov==4.0.0
pytest-django==4.5.2
pytest-xdist==3.0.2
django-q